Drop vStringAutoResize() and expose vStringResize()
This allows callers to directly allocate a string of the correct size
instead of reallocating strings many times. In addition, from the name
vStringAutoResize() it isn't really clear what it does (it just doubles
the string size). vStringResize() has been modified to always create
strings of lengths of the power 2 and not to reallocate strings which
already have the correct size.