docs, style: add dir-locals file for setting the style to linux in GNU emacs
Though .editorconfig was introduced, it is not enough for GNU emacs users to make
the coding style consistent with existing code of ctags.

The default style of cc-mode of GNU emacs is "gnu". The "gnu" style adds
extra indentation in block opening:

________if (...)
__________{

The style is not consistent with the existing code.
This commit introduces .dir-locals.el file which forces the style
to "linux". With "linux" style the block opening is handled as follows:

________if (...)
________{

The difficulties is that the basic indent size of "linux" is 8.
It is different from .editorconfig.

So in the updated document, I wrote the way to apply .editorconfig
configuration after installing "linux" style.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>