read: Fix theoretical access to freed memory
`readLineRawWithNoSeek()` could get the pointer to the internal string
buffer before last modifying it, if reading the last line of the input
and that line has no terminating `\r` or `\n`.  Thus, if the last
modification to the string resulted in an allocation move (e.g. if
`realloc()` returned a different pointer), `readLineRawWithNoSeek()`
would return the pointer to the previous, now invalid, string buffer.