main: quote output file name before passing it to system(3) function
Following command line doesn't work:

      $ ctags -o 'a b' ...

because a shell lauched from system(3) deals a whitespace between 'a'
and 'b' as a separator. The output file name is passed to system(3)
to run external sort command.

This commit adds code to put double and single quoets around the output
file name before passing it to system(3).

The issue is reported by Lorenz Hipp <lhipp@idealbonn.de> in a private mail.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>