SQL: Ignore token after WITH and WITHOUT
The SQL parser tries to read a function definition when encountering
the FUNCTION keyword.  However, the keyword is used in some other
situations:

    /* https://www.postgresql.org/docs/current/static/sql-createcast.html */

    CREATE CAST (source_type AS target_type)
	WITH FUNCTION function_name (argument_type [, ...])
	[ AS ASSIGNMENT | AS IMPLICIT ]

    CREATE CAST (source_type AS target_type)
	WITHOUT FUNCTION
	[ AS ASSIGNMENT | AS IMPLICIT ]

FUNCTION keyword after WITH or WITHOUT keyword should not be handled as
a function definition, so skip it there.

The test case is taken from the issue submitted by @karlb.

Closes #1235.

Based on a patch from @masatake.