main: refactor getFileLanguage*
Introduce struct GetLanguageRequest to make the
intention of calling getFileLanguage* clear.

As far as I can remember, originally getFileLanguage
is for guessing the language for given input file.

Now it does three things,

1. just guessing: open an input file, guess the parser and close the file
2. opening and guessing: open an input file, guess the parser and return the input MIO
3. guessing for give MIO (opened file): guess the parser

From the name of functions, it was not obvious that which one is requested
from client code.

WARNING: parseFileWithMio never free given mio. The behavior is changed.
Signed-off-by: Masatake YAMATO <yamato@redhat.com>