main, lxpath: introduce xpathFileSpecs and xpathFileSpecCount fields to parserDefinition
The logic for selecting a xml based parser is hard coded in selectByDTD.
The logic should be replaced with declarative data defined in each xml based parser.

Newly introduced xpathFileSpec is the field where the declarative datum are stored.
selectByDTD may access the fields via getXpathFileSpec and select a parser based on
the datum.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>