sh: use isBashFunctionChar as default check_char func
Close #1261.

bash accepts many non isalnum characters as function identifiers even when
not using the `function` keyword.

If `function` keyword is used, ctags has a chance to switch check_char
var to a function specialized to bash. For the case `function` keyword
is not used, this commit change the default value of check_char var to
isBashFunctionChar.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>