tokeninfo API: add a parameter to tokenUnreadFull to control resetting scopeIndex field of token
From my experience developing ldscript and dtd parsers, scopeIndex field of tokenInfo struct
should be reset in tokenUnread. My understanding is that a parser may call tokenUnread when
the parser recognizes the change of context during parsing; the parser wants to read the
token later in new context.

In other hand, the scopeIndex is typical information about context. The parser may want to reset it when
unread.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>