Tmain, json-writer: add a case for testing escaping
When emitting a tag whose name includes a tab char, it is escaped in
ctags level in *ctags output*.

This is json output. So the escaping sohuld be done in jansson level.
For testing how tab chars in a name is handled in json output, input.1,
a man page is added as input.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>