main: call anonReset in main part
anonReset was desigend to be called in parse method of a parser
which use anonymous identifiers.

1. Parser developers can forget calling the function, and
2. Calling anonReset in parse method doesn't work well with promise API.

2. needs more explanation.

See following pseudo code for yacc parser.

    %{
            struct {
                    int x, y;
            } point;
    %}

    %%

    %%
            struct {
                    int x, y;
            } vector;

There are two areas of C. For running C parser on the areas, yacc parser
uses promise API. The promise API calls the parse method of C parser for each
area. It means anonReset called twice.

As the result, anonIds for point and vector cannot be distinguished.
Even though more than two C areas in an input file, anonReset for C
parser should be called once. In other words, anonReset should be
called for each input file, not for each area.

A static variable, parsersUsedInCurrentInput, is introducd for
tracking all parsers used in the current input. The ctags main part
calls anonReset for parsers tracked by the variable only once for each
input. The variable is cleared when newly input file is prepared.

Using ptrArray for parsersUsedInCurrentInput is suggested by @b4n.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>