Elm: Add support for Elm-Lang via optlib  (#1260)
elm: Add config and tests for elm-lang

Signed-off-by: Masatake YAMATO <yamato@redhat.com>

* Add --sort=no to command line when running the test case.

* elm: update news.rst and add win32 build

* elm: Switch to GPL

* elm: use "import" patterns for clearing scope stack

* elm: capture Y of "import X as Y" as renamed imported namespaces

Following [discussion](https://github.com/universal-ctags/ctags/pull/1260)
with @masatake we treat imported renamed modules as namespaces
with a suitable description.
In `import X as Y` X is not captured as a reference tag, tagging Y is still
useful. A user of an editor can jump to "import X as Y" line
from where Y is referred. The user can know Y is alternative
name of X.