main: release resource for anonGenerate explicitly
parsersUsedInCurrentInput is used for tracking parsers used in the current
input file.

Once parsersUsedInCurrentInput was allocated, it was reused in input
files; instead ptrArrayClear was called for each input
file. parsersUsedInCurrentInput was never released.

This commit gives the life cycle of parsersUsedInCurrentInput explicitly;
parsersUsedInCurrentInput is allocated before parsing new file and
released after parsing it.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>