main, refactor: rename truncateLine field of tagEntryInfo
For handing too long source input lines, Universal ctags
has been extend the way to trancate them. As the result
the field name, truncateLine in tagEntryInfo struct is
too ambiguous to know what happens if it is true.

This commit rename it to truncateLineAfterTag to represnet
the intent of the field.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>