main: introduce thinInputStream
New concept subparser/baseparser combination is planed.
This commit is preparation for the concept.

Nested input stream and promise API were introduced to run a guest
parser which parsers a narrowed area of the origin input file.

subparser/baseparser may abuse them.

A subparser does nothing proactively; it is a set of callback
functions and is called from a baseparser.

When ctags chooses a baseparser for parsing an input file,
there is no issue.

When ctags chooses a subparser for parsing an input file,
there is an issue which is solved by the change of this commit.

A subparser can be chosen by its langmap. However, the subparser
itself do nothing proactively. The subparser must run associated base
parser on the same input file.

My idea original is that using promise API for applying the baseparser
on the input file already opened for the subparser.

However, promise API expects that it is used AFTER parsing the
original input with a parser. In subparser/baseparser combination,
promise API is used BEFORE parsing the original input with a chosen
subparser.

`thinDepth' counter is introduced to File object.
When a promise is made with the special area (0, 0, 0, 0),
no new input stream is stacked for the given parser.
Instead the counter is incremented. The parser is run on the original
stream when the promise is forced. After parsing, the
the stream is not popped because the counter is incremented. Instead
of popping the stream, the counter is decremented.

Let's apply sub/base parsers scenario to this new mechanism.
The main part chooses a sub parser and opens an input stream.
The subparser makes a promise with the special area to run the
associated base parser on the input stream. The promise is forced
but just `thinDepth' counter is incremented to the input stream.
The base parser runs on the stream. After running the input stream
gets a chance to be popped, but is not popped because the `thinDepth'
counter is not zero.

Iniially `thinDepth' was implemented as a flag. However, I recognized
usecase that subparsers are nested. So it was implemented as a
counter.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>