main: initialize base parsers from subparsers
In top down parser choice, a subparser may call
its base parser. The subparser is initialized well
when it is chosen. However, the base parser didn't
have a chance to be initialized. With this patch,
a base parser is also initialized with its subparser
only when a subparser that expects the top down parser
choice.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>