main,refactor: introduce parserObject
parserDefinition data type is too large; it includes members only used
in the main and members specified by parsers. These private and public
members should be separated.

This commit introduces parserObject.
It is for owning private fields.
parserObject points a parserDefinition.
parserDefinition is for owning public fields.

This is the first step; currently there is no
private field in parserObject.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>