Cobol: introduce division kind
Suggested by @FrankCai in #1324.
Partially close #1324.

The names of divisions are defined in the language, not newly defined
in cobol source files. So I wonder whether divisions should be tagged
or not. However, sections are already tagged. "sections" are defined
in the language, too. About divisions, I decide do the as sections.

TODO: These tags can be used to make scope field. and generate full
qualified tags.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>