units: normalize anon names
Partially close #1348.

When building ctags at the build directory differing from the source directory,
"make units" fails.

"make units" in such cross building the path for test input file is different
from the path used when generating expected.tags.

The difference of paths for the same input file causes the difference of
anon names in tags file because the hash value of a path is used to generate
an anon name.

I added a filter function for normalizing anon names in tags file.
--_anonhash is used internally.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>