html: Make the parser more resistent to invalid input
In cases when the input file is misformated and contains too many open
tags which are not closed, the parser can recurse too deeply and the
stack can overflow.

Add a sanity check to make sure there are not too many nested elements
and refuse to recurse if the number is exceeded (the chosen number 1000
should be more than enough for any valid HTML and still several orders
of magnitude smaller than usual stack size).