main: switch language when calling makeTagEntry for flusphing corkQueue
Switching language by pushLanguage/popLanguage is needed when calling
makeTagEntry from a subparser. This is job of a base parser.

However, it was not enough. When flushing corkQueue makeTagEntry was
called in the base parser context even when flushing an entry
made in a subparser. The flushing should be done in the sub parser
context.

TODO: Getting the langauge id of subparser is very slow. Ideally
an entry should hold the langauge id intead of its name.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>