ITcl: improve the way to handle methods and the other objects defined in a class
ITcl assumed a method is defiend at the top level.
It seems that this is wrong. A method is defined in a class.

With this commit ITcl parser captures methods defined in a class;
"method" used in the top level is ignored.

In addition variables and commons defined in a class are captured.

TODO: A separator used between a ITcl class and Tcl namespace should
be :: but . is used. This is a bug.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>