win32: Classify the source files properly on Visual Studio
On Visual Studio, files are listed under virtual folders which are
different from actual directory structure.  The folders are written in
*.vcxproj.filters files.

E.g.:

```
    <ClCompile Include="..\fnmatch\fnmatch.c">
      <Filter>Source Files\fnmatch</Filter>
    </ClCompile>
    <ClCompile Include="mkstemp\mkstemp.c">
      <Filter>Source Files\win32\mkstemp</Filter>
    </ClCompile>
```

This will be shown as the following folder structure:

```
    Source Files\
        fnmatch\
	    fnmatch.c
	win32\
	    mkstemp\
	        mkstemp.c
```

The `insert_final_newline` setting in .editorconfig is also changed,
because Visual Studio removes the final newline from project files.