[INCOMPATIBLE] main, Ruby, Rust, SQL, ObjectiveC: don't allow to use 'F' as a kind letter in parser
The kind letter 'F' is reserved  for file kind.
Universal-ctags provided the way to override 'F' for using the kind letter for the
parser own purpose. However, it will cause a trouble.

Though INCOMPATIBLE should not be introduced in Universal-ctags, however, keeping
constancy of tags file is more important than such compatibilities.
The initial release of Universal-ctags may be the last chance for us to
introduce such incompatible change.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>