[SELF INCOMPATIBLE] main: rename --_list-roles option to --list-roles
Signed-off-by: Masatake YAMATO <yamato@redhat.com>

Tmain: use all.d instead of all.h to test the output of --list-roles

A test case for the output of --list-roles is rearranged.
Glade parser uses 'h' as a name of role. Glade parser is available only
on the platform where libxml2 is installed. As the result the case
for testing --list-roles is not passed on compiler=cygwin platform.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>