main: re-adjust the EOL characters when _multiline regex long flag is specified
The byte offset is important to get correct result when _multiline regex long flag
is specified.

The original code converts '\r\n' at the end of line(EOL) to '\n' when
building allLines buffer that the regex engine for multiline pattern
scans.  The conversion makes the buffer short. As the result ctags
reports wrong patterns and line numbers (#1437).

This change re-converts the EOL characters to adjust the length of
line.

Close #1437.