main,sandbox: don't check the existence of input file if the input file is already opened
Before trying tasteLanguage(), the existence of the input file was checked always.
However, if the input file is already opened, this checking is just wasteful.

In the sandbox, the input is not specified with a file name, instead a mio
connected to stdin is given. So as the side effect this change

Close #1480

.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>