json-writer: decref an object passed to a json_pack explicitly
There were memory leaks of json objects passed to json_pack.  It seems
that I misunderstood the meaning of 'o' format chars of json_pack.
'o' was used and I expected that json_pack STEALS the ownership of the
passed object. However, valgrind detects the leak where 'o' is used.

This commits change where 'o' is used to 'O'.  With 'O', json_pack
works as usual; it takes the ownership of the passed object with
calling json_incref.

The memory leak is detected by a script based on q.sh in #1498
opened by @hanwen.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>