main,refactor: refactor the common statements in processKindsOption
This change the behavior of --<LANG>-kind option a bit.
When the control enters where warnings are printed, processKindsOption
returns false that means the passed option is not handled in
processKindsOption. In other words, broken option is not
consumed in processKindsOption.

This change makes processKindsOption consume broken options and
ignore them.

I think this is enough small incompatibility.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>