lregex: use the start of matched group when getting the line number for a tag
This is a bug fix of multi-line regex parser.

In the multi-line regex parser, the file offset for a tag is used as a
key for getting the line number for the tag. The offset comes from
a group of regex pattern; a parser specify the group number (here after
N) with a regex regex pattern.

The start of Nth group should be used to get the offset.  However, the
original code uses the end of Nth group.  As the result, unexpected
patterns and line numbers were recorded in tags file.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>