main,lregex: fill end: field of parent at the end of input
end: fields of parent entries can be filled if patterns with pop or
clear actions are matched. As side effects of the actions,
end: fields of entries in scope stacks are filled.

However, if input reaches EOF, these actions are never takens.
The end: fields of entries in scope stacks don't have a chance
to be filled.

This change forces to take the action equivalent to clear when
the input reaches EOF.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>