Fixed build script for BSD systems
In autogen.sh, don't assume that the correct Makefile interpreter is
`make`. The syntax of the Makefile generated by autotools is not
compatible with BSD Make, and must be built with GNU Make (gmake)
instead.

In addition to detecting BSD platforms (via uname) and using gmake
instead of make, autogen.sh also generates a BSDmakefile in the
top-level directory. In its presence, BSD Make will use it instead of a
generic Makefile, and BSD users will be informed that they should use
gmake instead of make to buil the project.

Tested on FreeBSD.