build-sys: don't use GNU make extension
It seems that binding a value to a macro in a target is GNU Make
extension. Our makefile uses the extension for setting a value to
TIMEOUT macro. However, it causes bmake reporting an error.

To use bmake as an build tool, this change removes lines using GNU
make extension.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>