go: put receiver type of function to scope
Following changes by @techee are merged in this commit:
simplify parseReceiver() a little

There should be no functional difference for valid code. The difference
in unit tests is for an invalid receiver and comes from the fact that
the code now works with the token passed from the caller and takes into
account the very first token too.

Simplyfy the condtion about receiver_type_token local variable.
Suggsted by @techee.