lregex: extend matchCallbackPattern about consuming input lines
If {exclusive} flag is specified and an input line matches a regex
pattern, the input line is consumed always; lregex skips the rest of
patterns for the input line. This is the original behavior of lregex
engine.

This commit extends the behavior if a pattern is specified with
matchCallbackPattern function. Even if a regex pattern (with
{exclusive}) matches an input line, and a callback false returns
false, the input line is never consumed. The callback function must
return true for consuming the input line.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>