jscript: Fix output encoding consistency with Unicode escape sequences
The input data read by the parsers is not already converted to the
output encoding, so we should emit the expanded Unicode sequences in
the input encoding so that it's consistent with an equivalent literal.

It is kind of incorrect in the end because the tag names are not
converted either at a later point, so the effective output might have
mixed encoding, but that's a separate issue.