Tmain: add case for testing emitting a tag with disabled kind (and role)
Some parsers check whether a kind is disabled or enabled when capturing a
tag or emitting a tag with makeTagEntry.

ctags allows a parser checking such thing. Even if a parser doesn't
check it, ctags should work well. It means such check should be done
in the main part. A parser can do it but is for optimizing.

This test case verifies the main part checks the
state(enabled or disabled) of a kind well.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>