go: Improve parsing of anonymous struct members
This patch tries to more clearly separate parsing of anonymous members
from parsing of ordinary non-anonymous struct members. The goal is that
if in the future type is recorded for non-anonymous struct members,
it is easy to collect the type.

The patch also tries to improve readability of the code by making things
more explicit, using better variable names (especially not using
memberCandidate for the anonymous member case) and adding more comments.