buildsys: Handle the case that `lib' is given as prefix for iconv functions
Close #1606.

On MacOSX, there is a case that lib is given as a prefix a
prefix for iconv functions.

The complexity comes from the fact that two libraries can be installed
in the case: one without the prefix and the other with the prefix.
Which one is linked to is up to -L options.

-L options are decided in configure. e.g. If libxml2, used in ctags,
is installed somewhere, configure adds a -L option. If libxml2 is not
installed anything, configure doesn't add.

We can known what kind of -L options are set to *_LIBS Makefile
variables at the end of configure. That means we can known whether
`lib' prefix is needed or not for using iconv at the end of configure.

This change introduces `ICONV_USE_LIB_PREFIX' C preprocessor Macro.
The detection the need of the prefix is reflected to the macro.