OCaml: remove unused kinds
"match" and "beginEnd" kinds were introduced accidentally.
The kind is refereed when making a tag entry in prepareTag() via calling
contextDescription.

contextDescription() converts the type of context pushed on `stack' to
kind index. The kind index is refereed as a value representing parent
scope when making a tag entry.

A parent scope must have a name. getLastNamedIndex() finds a NAMED
parent context. Such context is converted by contextDescription().

A context having `ContextMatch' or `ContextBlock' as its type has
no name; when calling a function pushing a context having such type to `stack',
NULL is passed as the context name.

    ContextMatch:

	    case OcaKEYWORD_function:
		    toDoNext = &matchPattern;
		    pushSoftContext (&matchPattern, NULL, ContextMatch);
		    break;

	    case OcaKEYWORD_with:
		    popSoftContext ();
		    toDoNext = &matchPattern;
		    pushSoftContext (&matchPattern, NULL, ContextMatch);
		    break;
    ContextBlock:

	    case OcaKEYWORD_begin:
		    pushContext (ContextStrong, ContextBlock, &mayRedeclare, NULL);
		    toDoNext = &mayRedeclare;
		    break;

A context having `ContextMatch' or `ContextBlock' as its type is never converted
to a kind index with contextDescription() because the context doesn't have a name.

That means "match" and "beginEnd" kinds can be removed safely, and we should remove
them.

These two strange kinds were introduced when I changed the representation of kind
from string to int. I didn't take enough time to think the detail of OCaml parser
because the code change was so large.

This change has no impact on the behavior of the parser. Both kinds
have never been used. Its related codes are not executed. The output
of --list-kinds=OCaml is just simplified.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>