[SELF INCOMPATIBLE] main: reserver '{' in kind description as a meta character
It was allowed to use any character including
`{' in the description part of --kinddef-<LANG> flag. In the future,
the option takes flags. In that case ctags detects `{' as the start
maker of the flags. To use `{' in the description, put '\' ahead of
`{' for escaping. This change handles `{' in the description as the
start marker of the flags, and `\' as the escape character.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>