main,refactor: don't check twice whether a tag is writable or not
Checking whether a tag was writable or not is conducted at the head of
both writeTagEntry and makeTagEntry.

This change simplifies the condition of checking; and reduces the
conditions.

If the cork is enabled, checking is done when writing a tag (writeTagEntry).
If not, checking is donw when accepting a tag (makeTagEntry).

Signed-off-by: Masatake YAMATO <yamato@redhat.com>