mtable|mline-lregex avoid running a parser twice for the same input file
When ctags reaches the end of the input file, ctags runs mtable and
mline parsers.

Reaching the end can happens twice. As the result ctags may run a
mtable (or mline) parser twice for the same input file. This is a bug.

See createTagsWithFallback1.  It calls createTagsForFile. During
running this function, reaching the end can happens once.

After calling createTagsForFile, createTagsWithFallback1 calls
readLineFromInputFile in a loop. Reaching the end can happens once
again.

To avoid the bug, we need a mechanism to limit running the parsers for
the same input only once.

This commit uses File.allLines as a flag.  ctags verifies
File.allLines is not NULL before running the parsers.  If it is NULL,
skip the running.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>