mtable-lregex,Tmain: limit a parser staying at the same input position so long time
By nature, a mtable regex parser can enter an infinite loop.
This change is for avoiding entering infinite loop resulted from a
simple programming mistake.

When a parser doesn't consume an input stream so long time, we can
assume it enters an infinite loop. In such cases, ctags stops running
the parser for the input with warning.

This commit deals with the following case:

A. a parser doesn't advance the input position repeatedly, and
B. its table stack doesn't grow so deep but it switches the tables.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>