main,Cxx,jscript: introduce "anonymous" extra
In ctags.1 man page, I wrote what is extras.

   Extras BASICALLY, ctags tags only language objects appeared on
       source files as is. In other words, a value for a name: field
       should be found on the source file associated with the name:.
       An extra tag (extra) is for tagging a language object with
       processed name or for tagging not associated with a language
       object. Typical extra tag is "qualified". That tags a language
       object with class-qualified or scope-qualified name.

An anonymous language object like lambda function doesn't appear on
source files. So it should be tagged as an extra tag.a

"anonymous" extra introduced in this commit is enabled by default.
So the impact on the existing users is small.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>