[SELF INCOMPATIBLE]Ctags,Elm,Gdbinit,Man,Passwd,Pod,PuppetManifest: Capitalize the names of optlib parsers
I would like to introduce a convention about naming parsers.  For a
built-in parser whether it is a crafted parser (parser written in C) or
an optlib parser (translated to C in build time), its name should be
capitalized; the name should be started from an uppercase letter.

This convention helps a user who wants to write a private optlib
parser; such a parser should have a name started from a lowercase
letter. As the result, we can avoid a name confliction between the
built-in parsers and the private optlib parsers.

This commit is for built-in optlib parsers.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>