Java,JSON,TTCN: use "boolean" instead of "bool"
All "boolean" were replaced with "bool" in
the commit "Prefer stdbool.h over mybool.h" (ce990805).

String literals for recognizing "boolean" keywords in Java and TTCN
parsers were also replaced. This should not be done.

About JSON parser, "boolean" was used as a name for a kind.  Changing
back from "bool" to "boolean" is not must. However, "boolean" may be
more popular word than "bool" for JSON users to represent a boolean
object. Actually geany'c json.c chooses "boolean" as the kind name.

(This change is not needed to be ported to geany's ctags.
 geany uses "boolean" correctly in the all places.)

Signed-off-by: Masatake YAMATO <yamato@redhat.com>