Tmain: initialize "V" variable explicitly
"V" variable was introduced as a placeholder that is used when I want
to run the test case under valgrind; the valgrind command line is set
to "V".

It seems that "Ubuntu PPA build" environment defines V as an environment
variable with value "1". See #1761.

As the result, the test case doesn't work well on the environment.

This commit clears the value for "V" to allow the test case works
on the environment.