[INCOMPATIBLE][SELF-INCOMPATIBLE] main,docs: don't allow to use numerical char as the first letter in a kind name
SELF-INCOMPATIBLE aspect:
When defining a kind with --kinddef-<LANG>, Universal-ctags accepted
both an alphabetical and a numerical characters as the initial letter.
With this change, Universal-ctags accepts only an alphabetical.

INCOMPATIBLE aspect:
Universal-ctags accepts numerical characters as the rest letters
(other than initial letter) in a kind name. This violates the tags v2
format. This commit documents it as an EXCEPTION.
We must convert the EXCEPTION to planed v3 format.

Exuberant-ctags doesn't have --kinddef-<LANG>.
However, a kind can be defined with --regex-<LANG>. This commit doesn't deal with
the option.

Re-wording is suggsted by @ploxiln.
Revising the error message is suggested by @ploxiln.
A wrong sentence is pointed out by @ploxiln.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>