[INCOMPATIBLE] main,docs(man),Tmain: allow to use alphabetical char in kind name defined with --regex-<LANG>
This change has big impact.
As far as browsing .ctags files in github, people users
whitespace in kind names defined with --regex-<LANG>.

This change rejects them.

A kind name is used for "value" part of tagfield.
In this case, there is no trouble in using
a whitespace (and characters other than alphabetical char).

However, a kind name can be used "name" pat of tagfield
when printing scope field. In this case, using characters
other than alphabetical char in the kind can be a cause of
trouble.

(After long thinking I decide to add an exception; using a numerical
char as parts of the body of a kind name is allowed. Here "body" means
non-initial letters of the kind name.)

Revising the error message is suggested by @ploxiln.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>