[SELF-INCOMPATIBLE] xformat-writer,json-writer: make typeref: field in xref and json output formats as same as that in tags
typeref: field has sub-fields "typeref prefix" and "typeref name" in tags format.
":" combines the two sub-fields and makes the field value.

e.g.

	typename:int
	struct:S

In xref and json output formats, only "typeref name"
sub field was printed when printing typeref: field.

This change fixes this inconsistent.
This is self-incompatible change.

If you want, "typerefPrefix" and "typerefName" fields can be introduced.
Open an issue in that case.

This change is needed to solve #1772.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>