Python: fix wrong role assignments
When making a reference tag of "unknown" kind,
PYTHON_MODULE_IMPORTED is used to assign a role for the tag.
PYTHON_UNKNOWN_IMPORTED should be used instead.

Our test cases could not detect the mistake because, The integer
representation of PYTHON_UNKNOWN_IMPORTED role enumerator and that of
PYTHON_MODULE_IMPORTED role enumerator are the same (== 1). In other
hand the string representation of the roles are the same (==
"imported"), too.

The API for defining and assigning roles must be improved in the
future.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>