autoit: Simplify skipping the matched keyword
Instead of leaving up to the caller to compute the position after a
successful match, add an output parameter to `match()` for returning
that position.

This makes the code more robust by not requiring the caller to
remember or compute how much to skip in case of a successful match,
which could easily lead to hard to track bugs in case the match and
skip length gets desynchronized.