main: allow a host parser to rearrange input passed to its guest parser
This change is about preprocessor.

ctags captures definitions in an input based on the syntax of the
preprocessor language.

The parser of preprocessor language may want to run another parser for
the input.

An simple example is CPreProcessor and C.

CPreProcessor parsers parse input.c to capture macros defined with #define.
Then CPreProcessor runs a parser for C language.

CPreProcessor parser is a host parser, and C parser is a guest parser here.

Just parsing the same input with a host parser and its guest parser
doesn't make good results because the guest parser may fall into the
unknown syntax of the preprocessor language.

With this change, a host parser gets a chance to arrange input before
running its guest parser. A host parser can delete lines about
directives to the preprocessor.

There are many limits in this change. A host parser can overwrite the
input but it cannot change the length of input lines.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>