main: separate the language stack from inputFile structure
The language stack was embedded in inputFile struct.

The design of memory managements for inputFile and the language stack
is different.

The inputFile can be stacked only 2 levels. The levels are represented
with BackupFile and File variables having the inputFile struct as
their type.  The value of the File is copied to the BackupFile when
pushing new input stream.  It is done reversely when popping. The code
managing inputFile doesn't allocate memory dynamically.

In the other hand, the language stack can grows any levels of
depth. It uses realloc internally to extend the depth.

There was a case that a pointer for dynamically allocate memory for
extending the language stack is overwritten when copying
for the inputFile level management. This causes a crash.

The crash can be avoided with following patch:

    diff --git a/main/read.c b/main/read.c
    index 593aeac1..81d1c81a 100644
    --- a/main/read.c
    +++ b/main/read.c
    @@ -1205,7 +1205,11 @@ extern void   popNarrowedInputStream  (void)
		    return;
	    }
	    mio_free (File.mio);
    +       langType *tmp_languages = File.input.langInfo.stack.languages;
    +       unsigned int tmp_size = File.input.langInfo.stack.size;
	    File = BackupFile;
    +       File.input.langInfo.stack.languages = tmp_languages;
    +       File.input.langInfo.stack.size = tmp_size;
	    memset (&BackupFile, 0, sizeof (BackupFile));
     }

However, the above change increases the complexity of code.  Instead,
this change separates the language stack from the inputFile structure.
As the result copying based the stack manipulation for the inputFile
structure and that for the language stack can be done independently.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>