vstring: Use memcpy() instead of strncpy()
We don't need the extra strncpy() semantic in stringCat() as we always
know the length we want, so it's slightly faster to use memcpy(), about
2% in some cases.

Additionally, this would improve support for NUL bytes embedded in a
vString, avoiding an extra truncation.  However, no consumer currently
rely on NUL byte support so it currently has no impact on this end.